minor organizational changes, compare Map test image
