+++ .gitignore
@@ -1,6 +1,5 @@
-/build
-/lib
-/node_modules
-/test/actual
-/test/results
-image.png
+build
+lib
+node_modules
+test/actual
+test/diff
