+++ test/expected/gzip/success.png
