+++ test/expected/gzip/unhandled.png
